[ALPS07689827] kernel-6.0: Fix build error
Kernel-6.0 has stricter error checks for printf and function declaration

error: data argument not used by format string [-Werror,-Wformat-extra-args]
error: more % conversions than data arguments [-Werror,-Wformat-insufficient-args]
error: type specifier missing, defaults to int [-Werror,-Wimplicit-int]

MTK-Commit-Id: 8db5d5aeb14183365067caa5b2bfc1085490a8f5

CR-Id: ALPS07689827
Feature: Kernel SI Operation

+++ drivers/media/platform/mtk-isp/camsys/isp7_1/cam/mtk_cam-mraw.c
@@ -1613,7 +1613,7 @@ int mtk_cam_mraw_cq_disable(struct mtk_mraw_device *dev)
 {
 	int ret = 0;
 
-	writel_relaxed(~CQ_THR0_EN, dev->base + REG_MRAW_CQ_THR0_CTL);
+	writel_relaxed((u32)(~CQ_THR0_EN), dev->base + REG_MRAW_CQ_THR0_CTL);
 	wmb(); /* TBC */
 
 	return ret;
